Use annotations in errorprone Motivation: There are some useful annotations in errorprone.
Modifications:
- Use `@FormatMethod` with `@FormatString` and `@MustBeClosed` where applicable
- Replace `@CheckReturnValue` and `@GuardedBy` with the ones in errorprone
- Remove `@NotThreadSafe` that doesn't mean much

Result:
- More validations

+++ core/src/test/java/com/linecorp/armeria/server/TestConverters.java
@@ -38,7 +38,7 @@ public final class TestConverters {
                                             @Nullable Object result,
                                             HttpHeaders trailers) throws Exception {
             if (result instanceof Integer) {
-                return httpResponse(HttpData.ofUtf8("Integer: %d", result));
+                return httpResponse(HttpData.ofUtf8("Integer: %d", (Integer) result));
             }
             return ResponseConverterFunction.fallthrough();
         }
